requirements.txt
requests
python-dotenv
beautifulsoup4
lxml
psycopg[binary]
pymupdf
pdf2image
pytesseract
groq
openai